remove the ability to delete own user account Запретил удалять свою учётную запись

+++ routers/web/user/setting/account.go
@@ -247,7 +247,9 @@ func DeleteAccount(ctx *context.Context) {
 		}
 		return
 	}
-
+	if ctx.Doer.IsAdmin == false {
+		return
+	}
 	if err := user.DeleteUser(ctx, ctx.Doer, false); err != nil {
 		switch {
 		case models.IsErrUserOwnRepos(err):
